Fix a few bash-related issues:
1. Bring in a hookbook update for better DEBUG hook handling.
2. Work around bash's inability to `x | source /dev/stdin`.

For the hookbook change, see https://github.com/Shopify/hookbook/pull/2

As far as (2): zsh is perfectly happy to `x | source /dev/stdin`, but
bash barfs on this in many situations (relating, I think, to the data
size).

To fix this, we `eval "$(x)"` instead, but that means that shadowenv
can't just take its PPID and have that be meaningful. Instead, we pass
the shell's PID explicitly to shadowenv.