Handle unset and empty properly.
The list of items in the path list was improperly being set to a list
containing empty string ([""]) regardless of whether or not the path
list was previously set. This change fixes the handling of unset path
lists.

Before:

nil        =>  [""]
""         =>  [""]
"foo:bar"  =>  ["foo", "bar"]

After:

nil        =>  []
""         =>  [""]
"foo:bar"  =>  ["foo", "bar"]

Co-authored-by: Lisa Ugray <lisa.ugray@shopify.com>