Implement `shadowenv exec --dir <dir>`:
This causes shadowenv exec to load the shadowenv from a given directory,
rather than the current directory, but still execute the command
provided from the current directory after loading that other
environment.

I've also removed the short form (-d) for shadowenv exec
--shadowenv-data, since it just became very confusing. I don't think
anyone was really using that, so removing it should be ok.