Fix key loading issue:
When we upgraded dalek-ed25519, we switched from a Keypair-centric API
to a SigningKey-centric API, but inadvertently changed the format of the
generated trust-key-v2 file to only include the private key.

This is actually a better format, so we'll keep using it, and just
truncate an incoming 64-byte file to 32 bytes to get the signing key, if
that's the format it has.