Don't explode when expand-path result is missing
Expressions like (expand-path "~/does-not-exist") explode since
std::fs::canonicalize needs to point to a file that exists. In the vast
majority of cases this isn't desirable, nor is its behaviour of
resolving symlinks. Symlinks subtly break shadowenv's eval caching
mechanism because the value does not get updated if the symlink changes,
and shadowenv explodes if the there is a symlink loop. If people want
that behaviour for some reason I think we should expose it via a
different expression, or ideally people should canonicalize the paths
externally.