Don't explode when expand-path result is missing
Expressions like (expand-path "~/does-not-exist") explode since
std::fs::canonicalize needs to point to a file that exists. In the vast
majority of cases this isn't desirable, nor is its behaviour of
resolving symlinks. If people want that for some reason I think we
should expose that via a different expression.