Improve handling of duplicate paths
Currently if we add a duplicate path to a pathlist, it won't get removed
when the changes are unwound. By taking the number of occurrences into
account we can at least remove the duplicate that we added correctly. We
don't have enough info to know which occurence to remove so we always
remove the ones at the front of the pathlist since that's the most
common case.