Update documentation for `shadowenv exec`
This change updates the documentation to be more explicit about where to
put command and arguments.

The previous documentation suggested that program arguments need to be
specified after a `--`. But it's actually both the command name and the
arguments.

The old text:

$ shadowenv exec --help
...
USAGE:
    shadowenv exec [OPTIONS] <child-argv0|child-argv> [child-argv0] [-- <child-argv>...]

ARGS:
    <child-argv0>      If the command takes no arguments, it can be passed directly as the last arugment.
    <child-argv>...    If the command requires arguments, they must all be passed after a --.

This is in contrast of how it actually works.

Bad:

$ shadowenv exec ls -- /
error: The argument '<child-argv>...' cannot be used with one or more of
  the other specified arguments

Good:

$ shadowenv exec -- ls /
Applications    cores           etc ...